refactor: extract MainView UserControl from MainWindow shell
